+++ src_plugins/ppx_deriving_create.ml
@@ -5,18 +5,18 @@ open Parsetree
 open Ast_helper
 open Ast_convenience
 
-let prefix = "create"
+let deriver = "create"
 let raise_errorf = Ppx_deriving.raise_errorf
 
 let find_main labels =
   List.fold_left (fun (main, labels) ({ pld_type; pld_loc } as label) ->
-    match Ppx_deriving.attr ~prefix "main" pld_type.ptyp_attributes with
+    match Ppx_deriving.attr ~deriver "main" pld_type.ptyp_attributes with
     | Some (_, PStr []) ->
       begin match main with
-      | Some _ -> raise_errorf ~loc:pld_loc "Duplicate [@deriving.%s.main] annotation" prefix
+      | Some _ -> raise_errorf ~loc:pld_loc "Duplicate [@deriving.%s.main] annotation" deriver
       | None -> Some label, labels
       end
-    | Some ({ loc }, _) -> raise_errorf ~loc "Invalid [@deriving.%s.main] syntax" prefix
+    | Some ({ loc }, _) -> raise_errorf ~loc "Invalid [@deriving.%s.main] syntax" deriver
     | None -> main, label :: labels)
     (None, []) labels
 
@@ -36,12 +36,12 @@ let str_of_type ~options ~path ({ ptype_loc = loc } as type_decl) =
           Exp.fun_ "" None (punit ()) (record fields)
       in
       List.fold_left (fun accum { pld_name = { txt = name }; pld_type } ->
-        match Ppx_deriving.attr ~prefix "default" pld_type.ptyp_attributes with
+        match Ppx_deriving.attr ~deriver "default" pld_type.ptyp_attributes with
         | Some (_, PStr [{ pstr_desc = Pstr_eval (default, _) }]) ->
           Exp.fun_ ("?"^name) (Some default) (pvar name) accum
-        | Some ({ loc }, _) -> raise_errorf ~loc "Invalid [@deriving.%s.default] syntax" prefix
+        | Some ({ loc }, _) -> raise_errorf ~loc "Invalid [@deriving.%s.default] syntax" deriver
         | None ->
-        match Ppx_deriving.attr ~prefix "split" pld_type.ptyp_attributes with
+        match Ppx_deriving.attr ~deriver "split" pld_type.ptyp_attributes with
         | Some (_, PStr []) ->
           begin match pld_type with
           | [%type: [%t? lhs] * [%t? rhs] list] when name.[String.length name - 1] = 's' ->
@@ -50,9 +50,9 @@ let str_of_type ~options ~path ({ ptype_loc = loc } as type_decl) =
               (Exp.fun_ ("?"^name) (Some [%expr []]) (pvar name)
                 [%expr let [%p pvar name] = [%e evar name'], [%e evar name] in [%e accum]])
           | _ -> raise_errorf ~loc "[@deriving.%s.split] annotation requires a type of form \
-                                    'a * 'b list and label name ending with `s'" prefix
+                                    'a * 'b list and label name ending with `s'" deriver
           end
-        | Some ({ loc }, _) -> raise_errorf ~loc "Invalid [@deriving.%s.split] syntax" prefix
+        | Some ({ loc }, _) -> raise_errorf ~loc "Invalid [@deriving.%s.split] syntax" deriver
         | None ->
         match pld_type with
         | [%type: [%t? _] list] ->
@@ -61,9 +61,9 @@ let str_of_type ~options ~path ({ ptype_loc = loc } as type_decl) =
           Exp.fun_ ("?"^name) None (pvar name) accum
         | _ -> Exp.fun_ name None (pvar name) accum)
         fn labels
-    | _ -> raise_errorf ~loc "Can only derive %s for record types" prefix
+    | _ -> raise_errorf ~loc "%s can be derived only for record types" deriver
   in
-  [Vb.mk (pvar (Ppx_deriving.mangle_type_decl (`Prefix prefix) type_decl)) mapper]
+  [Vb.mk (pvar (Ppx_deriving.mangle_type_decl (`Prefix deriver) type_decl)) mapper]
 
 let wrap_predef_option typ =
   let predef_option = mknoloc (Ldot (Lident "*predef*", "option")) in
@@ -83,11 +83,11 @@ let sig_of_type ~options ~path ({ ptype_loc = loc } as type_decl) =
           Typ.arrow "" (tconstr "unit" []) typ
       in
       List.fold_left (fun accum { pld_name = { txt = name; loc }; pld_type } ->
-        match Ppx_deriving.attr ~prefix "default" pld_type.ptyp_attributes with
+        match Ppx_deriving.attr ~deriver "default" pld_type.ptyp_attributes with
         | Some (_, PStr _)  -> Typ.arrow ("?"^name) (wrap_predef_option pld_type) accum
-        | Some ({ loc }, _) -> raise_errorf ~loc "Invalid [@deriving.%s.default] syntax" prefix
+        | Some ({ loc }, _) -> raise_errorf ~loc "Invalid [@deriving.%s.default] syntax" deriver
         | None ->
-        match Ppx_deriving.attr ~prefix "split" pld_type.ptyp_attributes with
+        match Ppx_deriving.attr ~deriver "split" pld_type.ptyp_attributes with
         | Some (_, PStr []) ->
           begin match pld_type with
           | [%type: [%t? lhs] * [%t? rhs] list] when name.[String.length name - 1] = 's' ->
@@ -96,9 +96,9 @@ let sig_of_type ~options ~path ({ ptype_loc = loc } as type_decl) =
               (Typ.arrow ("?"^name) (wrap_predef_option [%type: [%t rhs] list])
                 accum)
           | _ -> raise_errorf ~loc "[@deriving.%s.split] annotation requires a type of form \
-                                    'a * 'b list and label name ending with `s'" prefix
+                                    'a * 'b list and label name ending with `s'" deriver
           end
-        | Some ({ loc }, _) -> raise_errorf ~loc "Invalid [@deriving.%s.split] syntax" prefix
+        | Some ({ loc }, _) -> raise_errorf ~loc "Invalid [@deriving.%s.split] syntax" deriver
         | None ->
         match pld_type with
         | [%type: [%t? _] list] ->
@@ -107,12 +107,12 @@ let sig_of_type ~options ~path ({ ptype_loc = loc } as type_decl) =
           Typ.arrow ("?"^name) (wrap_predef_option opt) accum
         | _ -> Typ.arrow name pld_type accum)
         typ labels
-    | _ -> raise_errorf ~loc "Can only derive %s for record types" prefix
+    | _ -> raise_errorf ~loc "%s can only be derived for record types" deriver
   in
-  [Sig.value (Val.mk (mknoloc (Ppx_deriving.mangle_type_decl (`Prefix prefix) type_decl)) typ)]
+  [Sig.value (Val.mk (mknoloc (Ppx_deriving.mangle_type_decl (`Prefix deriver) type_decl)) typ)]
 
 let () =
-  Ppx_deriving.(register prefix {
+  Ppx_deriving.(register deriver {
     core_type = None;
     structure = (fun ~options ~path type_decls ->
       [Str.value Nonrecursive (List.concat (List.map (str_of_type ~options ~path) type_decls))]);