+++ src_plugins/ppx_deriving_ord.ml
@@ -5,7 +5,7 @@ open Parsetree
 open Ast_helper
 open Ast_convenience
 
-let prefix = "ord"
+let deriver = "ord"
 let raise_errorf = Ppx_deriving.raise_errorf
 
 let argn kind =
@@ -27,9 +27,9 @@ let rec exprsn typs =
     app (expr_of_typ typ) [evar (argn `lhs i); evar (argn `rhs i)])
 
 and expr_of_typ typ =
-  match Ppx_deriving.attr ~prefix "compare" typ.ptyp_attributes with
+  match Ppx_deriving.attr ~deriver "compare" typ.ptyp_attributes with
   | Some (_, PStr [{ pstr_desc = Pstr_eval (equal, _) }]) -> equal
-  | Some ({ loc }, _) -> raise_errorf ~loc "Invalid [@deriving.%s.compare] syntax" prefix
+  | Some ({ loc }, _) -> raise_errorf ~loc "Invalid [@deriving.%s.compare] syntax" deriver
   | None ->
     match typ with
     | [%type: int] | [%type: int32] | [%type: Int32.t]
@@ -83,8 +83,8 @@ and expr_of_typ typ =
             Exp.case (pdup (fun var -> Pat.alias (Pat.type_ tname) (mknoloc var)))
                      (app (expr_of_typ typ) [evar "lhs"; evar "rhs"])
           | _ ->
-            raise_errorf ~loc:ptyp_loc "Cannot derive ord for %s"
-                         (Ppx_deriving.string_of_core_type typ))
+            raise_errorf ~loc:ptyp_loc "%s cannot be derived for %s"
+                         deriver (Ppx_deriving.string_of_core_type typ))
       in
       let int_cases =
         fields |> List.mapi (fun i field ->
@@ -102,8 +102,8 @@ and expr_of_typ typ =
     | { ptyp_desc = Ptyp_var name } -> evar ("poly_"^name)
     | { ptyp_desc = Ptyp_alias (typ, _) } -> expr_of_typ typ
     | { ptyp_loc } ->
-      raise_errorf ~loc:ptyp_loc "Cannot derive ord for %s"
-                   (Ppx_deriving.string_of_core_type typ)
+      raise_errorf ~loc:ptyp_loc "%s cannot be derived for %s"
+                   deriver (Ppx_deriving.string_of_core_type typ)
 
 let str_of_type ~options ~path ({ ptype_loc = loc } as type_decl) =
   let comparator =
@@ -131,9 +131,9 @@ let str_of_type ~options ~path ({ ptype_loc = loc } as type_decl) =
       in
       [%expr fun lhs rhs -> [%e List.fold_left compare_reduce [%expr 0] exprs]]
     | Ptype_abstract, None ->
-      raise_errorf ~loc "Cannot derive ord for fully abstract type"
+      raise_errorf ~loc "%s cannot be derived for fully abstract types" deriver
     | Ptype_open, _ ->
-      raise_errorf ~loc "Cannot derive ord for open type"
+      raise_errorf ~loc "%s cannot be derived for open types" deriver
   in
   let polymorphize = Ppx_deriving.poly_fun_of_type_decl type_decl in
   [Vb.mk (pvar (Ppx_deriving.mangle_type_decl (`Prefix "compare") type_decl))
@@ -147,7 +147,7 @@ let sig_of_type ~options ~path type_decl =
               (polymorphize [%type: [%t typ] -> [%t typ] -> int]))]
 
 let () =
-  Ppx_deriving.(register "ord" {
+  Ppx_deriving.(register deriver {
     core_type = Some expr_of_typ;
     structure = (fun ~options ~path type_decls ->
       [Str.value Recursive (List.concat (List.map (str_of_type ~options ~path) type_decls))]);