+++ src_plugins/ppx_deriving_show.ml
@@ -5,20 +5,20 @@ open Parsetree
 open Ast_helper
 open Ast_convenience
 
-let prefix = "show"
+let deriver = "show"
 let raise_errorf = Ppx_deriving.raise_errorf
 
 let argn = Printf.sprintf "a%d"
 
 let rec expr_of_typ typ =
-  match Ppx_deriving.attr ~prefix "printer" typ.ptyp_attributes with
+  match Ppx_deriving.attr ~deriver "printer" typ.ptyp_attributes with
   | Some (_, PStr [{ pstr_desc = Pstr_eval (printer, _) }]) ->
     [%expr (let fprintf = Format.fprintf in [%e printer]) fmt [@ocaml.warning "-26"]]
-  | Some ({ loc }, _) -> raise_errorf ~loc "Invalid [@deriving.%s.printer] syntax" prefix
+  | Some ({ loc }, _) -> raise_errorf ~loc "Invalid [@deriving.%s.printer] syntax" deriver
   | None ->
-  match Ppx_deriving.attr ~prefix "opaque" typ.ptyp_attributes with
+  match Ppx_deriving.attr ~deriver "opaque" typ.ptyp_attributes with
   | Some (_, PStr []) -> [%expr fun _ -> Format.pp_print_string fmt "<opaque>"]
-  | Some ({ loc }, _) -> raise_errorf ~loc "Invalid [@deriving.%s.opaque] syntax" prefix
+  | Some ({ loc }, _) -> raise_errorf ~loc "Invalid [@deriving.%s.opaque] syntax" deriver
   | None ->
     let format x = [%expr Format.fprintf fmt [%e str x]] in
     let seq start finish fold typ =
@@ -58,11 +58,11 @@ let rec expr_of_typ typ =
       [%expr fun _ -> Format.pp_print_string fmt "<fun>"]
     | { ptyp_desc = Ptyp_constr ({ txt = lid }, args) } ->
       let args_pp = List.map (fun typ -> [%expr fun fmt -> [%e expr_of_typ typ]]) args in
-      begin match Ppx_deriving.attr ~prefix "polyprinter" typ.ptyp_attributes with
+      begin match Ppx_deriving.attr ~deriver "polyprinter" typ.ptyp_attributes with
       | Some (_, PStr [{ pstr_desc = Pstr_eval (printer, _) }]) ->
         app [%expr (let fprintf = Format.fprintf in [%e printer]) [@ocaml.warning "-26"]]
             (args_pp @ [[%expr fmt]])
-      | Some ({ loc }, _) -> raise_errorf ~loc "Invalid [@deriving.%s.polyprinter] syntax" prefix
+      | Some ({ loc }, _) -> raise_errorf ~loc "Invalid [@deriving.%s.polyprinter] syntax" deriver
       | None ->
         app (Exp.ident (mknoloc (Ppx_deriving.mangle_lid (`Prefix "pp") lid)))
             (args_pp @ [[%expr fmt]])
@@ -91,15 +91,15 @@ let rec expr_of_typ typ =
             Exp.case [%pat? [%p Pat.type_ tname] as x]
                      [%expr [%e expr_of_typ typ] x]
           | _ ->
-            raise_errorf ~loc:ptyp_loc "Cannot derive show for %s"
-                         (Ppx_deriving.string_of_core_type typ))
+            raise_errorf ~loc:ptyp_loc "%s cannot be derived for %s"
+                         deriver (Ppx_deriving.string_of_core_type typ))
       in
       Exp.function_ cases
     | { ptyp_desc = Ptyp_var name } -> [%expr [%e evar ("poly_"^name)] fmt]
     | { ptyp_desc = Ptyp_alias (typ, _) } -> expr_of_typ typ
     | { ptyp_loc } ->
-      raise_errorf ~loc:ptyp_loc "Cannot derive show for %s"
-                   (Ppx_deriving.string_of_core_type typ)
+      raise_errorf ~loc:ptyp_loc "%s cannot be derived for %s"
+                   deriver (Ppx_deriving.string_of_core_type typ)
 
 let str_of_type ~options ~path ({ ptype_loc = loc } as type_decl) =
   let path = Ppx_deriving.path_of_type_decl ~path type_decl in
@@ -136,8 +136,10 @@ let str_of_type ~options ~path ({ ptype_loc = loc } as type_decl) =
         [%e fields |> Ppx_deriving.(fold_exprs
               (seq_reduce ~sep:[%expr Format.fprintf fmt ";@ "]))];
         Format.fprintf fmt "@] }"]
-    | Ptype_abstract, None -> raise_errorf ~loc "Cannot derive show for fully abstract type"
-    | Ptype_open, _        -> raise_errorf ~loc "Cannot derive show for open type"
+    | Ptype_abstract, None -> 
+      raise_errorf ~loc "%s cannot be derived for fully abstract types" deriver
+    | Ptype_open, _        -> 
+      raise_errorf ~loc "%s cannot be derived for open types" deriver
   in
   let pp_poly_apply = Ppx_deriving.poly_apply_of_type_decl type_decl (evar
                         (Ppx_deriving.mangle_type_decl (`Prefix "pp") type_decl)) in
@@ -158,7 +160,7 @@ let sig_of_type ~options ~path type_decl =
               (polymorphize [%type: [%t typ] -> string]))]
 
 let () =
-  Ppx_deriving.(register "show" {
+  Ppx_deriving.(register deriver {
     core_type = Some (fun typ ->
       [%expr fun x -> Format.asprintf "%a" (fun fmt -> [%e expr_of_typ typ]) x]);
     structure = (fun ~options ~path type_decls ->